Remove optionxform override
While this keeps the case of keys in the output config this had the
side-effect of breaking expected behavior of the configparser object.
Keys are, at the parent level, case insensitive. The written output
removing case from the keys is the same as indentation being lost. A
feature of the parent, not the child.