Rename and isolate injected state variables
Two class variables are injected outside of the `__init__()` of the
class. This is done, mostly, to avoid the complexity of matching the
parent class init just to declare these.

Name mangling these two will prevent any possibility of collision with
the parent class and feel like best practice here.

Renaming them to be more accurate to what they are.