Preserve blank line formatting
resolves #68

This change treats emtpy lines as comments allowing the preservation of
them from read to write.

The changes in #73 are removed here. #73 specifically corrected the
ending of the file. Instead, this change takes full control of empty
lines. This corrects the issue #73 solved as well as correcting extra
empty lines that were appearing between sections during development.