Preserve option case between read and write (#75)
Normally options are normalized to lowercase during the read process.
This results in a formatting change on write that can be frustrating if
camelCase is used for option names.

This change ensures the `optionxform()` method does not apply the
`.lower()` converstion during reading.

Options remain case insensitive while using a ConfigParser object as
expected.